# Fix the residual norm reduction criterion for CUDA devices