--- a/core/stop/residual_norm_reduction_kernels.hpp
+++ b/core/stop/residual_norm_reduction_kernels.hpp
@@ -53,7 +53,7 @@ namespace residual_norm_reduction {
         const matrix::Dense<_type> *tau, const matrix::Dense<_type> *orig_tau, \
         remove_complex<_type> rel_residual_goal, uint8 stoppingId,             \
         bool setFinalized, Array<stopping_status> *stop_status,                \
-        bool *all_converged, bool *one_changed)
+        Array<bool> *device_storage, bool *all_converged, bool *one_changed)
 
 
 #define DECLARE_ALL_AS_TEMPLATES  \