This MR does everything that can be done to accelerate the Residual Norm Reduction Stopping Criterion for GPUs. Multiple optimizations are done to ensure no temporary data are allocated on the GPU, and therefore no cudaFree (and related) are called.
	1. An array to store the residual norms is always created to accelerate the norm computation (it is used only in the case we have a GPU executor AND the solver doesn't provide residual norms, which is currently all of Ginkgo).
	2. A small temporary array of two bytes is always created (used on the GPU) to represent the following booleans: `all_converged`, `one_changed`. There, I directly call `executor->copy_from()` as I found that was the most fitting function for that purpose.


Currently, results show on a P100 that the total criterion check takes 130 us, roughly 60us of that is the norm computation. A SpMV (with CSR and thermal 2 matrix) is 367 us by comparison. On the P100, the total criterion check time was divided by two. When looking at the BiCGSTAB solver time, it got down to `0.75289 s` from `0.81687 s` with thermal2 in CSR format (roughly 8% speedup).


I think that it would be possible to accelerate this further (make it almost perfect, i.e. divide the time by two again) but the only way to do it would be to migrate all of our copies towards using the "Async" modes for CUDA and/or using different streams.

The traces are available there:
[before fix](https://drive.google.com/open?id=1F0v-uw-3BANulesJKaV5Rxsw91ncsa3k)
[after fix](https://drive.google.com/open?id=1qJiwVFzHiYQGwGVso9sFAiYzPTa1MI1y)