@tcojean 2018-11-28
I did some quick tests which used Asynchronous copies in Ginkgo, and this confirms that the final compression of the stopping criterion time happens in that case (the preconditionner copy also gets overlapped). The problem is that making such a mode work correctly needs more thinking: we need to add `cudaStreamWaitEvent()` functions (or equivalent using the new cuda graphs and dependencies) to ensure kernels wait for the correct memory copy, among other things. This should require a non trivial evolution of the CUDA executor (and/or operation) model.