@gflegar 2018-11-28
> The problem is that making such a mode work correctly needs more thinking: we need to add `cudaStreamWaitEvent()` functions (or equivalent using the new cuda graphs and dependencies) to ensure kernels wait for the correct memory copy, among other things. This should require a non trivial evolution of the CUDA executor (and/or operation) model.

Could you create an issue about that. I'll throw in some ideas there as soon as I find some time to do it.