@hartwiganzt 2018-11-28
@tcojean can you please add screenshots of the nvvp output? I know, you cannot catch everything with a screenshot, but it is easier to grasp the overall story.