@tcojean 2018-12-04
Here are the two images from the previous traces for reference. The relevant portion is the violet part (included, this is the norm computation) until the beginning of the blue part. As we can see, the white portion in between is shrinked but what is left is due to synchronisations and can only be shrinked by using asynchronous copies. I will create an issue on this topic later.

Before the MR:
![](https://i.postimg.cc/WD0FQ7g8/criterion-arrays-unfixed.png)

With this MR:
![](https://i.postimg.cc/FYMkW6d1/criterion-arrays-fixed.png)