@tcojean 2018-12-05
I improved the performance one last time by using a kernel for the initialization of the booleans on the GPU instead of a synchronous copy. I think for the moment this is the most performance we can get out of the current stopping criterion design.

Here is a trace obtained on the FineCI system (K20). One spmv is 1.2ms, a dot 85us.
![](https://i.postimg.cc/qMQr8CyX/criterion-array-lastfix.png)